+++ tests/core.spec.ts
@@ -8,12 +8,23 @@ const translationsMock = {
   }
 }
 
+const translationsMockWithVariables = {
+  greeting: (name: string) => ({
+    de_DE: `Hallo ${name}`,
+    en_US: `Hello ${name}`,
+    pl_PL: `Cześć ${name}`
+  })
+}
+
 describe('translate', () => {
     it('should be a function', () => {
       expect(typeof translate).toBe('function')
     })
     it("should translate flat translations", () => {
       expect(translate(translationsMock, "en_US").example).toBe("Example")
+    });
+    it("should translate flat translations with variables", () => {
+      expect(translate(translationsMockWithVariables, "en_US").greeting("Mike")).toBe("Hello Mike");
     });
     it("should throw error if translations is invalid type", () => {
       const fn = () => {