+++ tests/withLittera.spec.tsx
@@ -9,15 +9,25 @@ const mockTranslations = {
     de_DE: "Einfach",
     pl_PL: "Proste",
     en_US: "Simple"
-  }
+  },
+  hello: (name: string) => ({
+    de_DE: `Hallo ${name}`,
+    pl_PL: `Cześć ${name}`,
+    en_US: `Hello ${name}`
+  })
 };
 
 const mockTranslated = {
-  simple: "Simple"
+  simple: "Simple",
+  hello: `Hello Mike`
 };
 
-const Component = withLittera(mockTranslations)(({translated}: LitteraProps<any>) => {
+const Component = withLittera(mockTranslations)(({translated}) => {
   return <div>{translated.simple}</div>
+});
+
+const ComponentWithVar = withLittera(mockTranslations)(({translated}) => {
+  return <div>{translated.hello("Mike")}</div>
 });
 
 describe('withLittera', () => {
@@ -34,5 +44,16 @@ describe('withLittera', () => {
       });
 
     expect(wrapper.find("div").props().children).toBe(mockTranslated.simple);
+  });
+
+  it('should render the component with translated text with variables', () => {
+    const wrapper = mount(
+      <ComponentWithVar />, 
+      {
+        wrappingComponent: LitteraProvider, 
+        wrappingComponentProps: {locales: ["en_US"], initialLocale: "en_US"}
+      });
+
+    expect(wrapper.find("div").props().children).toBe(mockTranslated.hello);
   });
 })